Integrated swr in the department and teams components
