+++ lib/fetcher.ts
@@ -1,0 +1,5 @@
+export const fetcher = async (url: string) => {
+  return fetch(url).then((r) => {
+    return r.json();
+  });
+};
